requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
bidict==0.22.0
boto3==1.24.84
botocore==1.27.84
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cloudpickle==1.6.0
colorama==0.4.5
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==38.0.1
cycler==0.11.0
decorator==4.4.2
dm-tree==0.1.7
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
future==0.18.2
gast==0.3.3
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.49.1
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.20.4
httptools==0.5.0
httpx==0.15.4
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==5.2.0
mysql-connector-python==8.0.30
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
Pillow==9.2.0
prompt-toolkit==2.0.10
protobuf==3.19.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.5.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.18.1
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2021.3
questionary==1.9.0
rasa==2.8.2
rasa-sdk==2.8.6
redis==3.5.3
regex==2021.7.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sanic==20.12.7
Sanic-Cors==0.10.0.post3
sanic-jwt==1.8.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.9.1
sentry-sdk==1.2.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
sniffio==1.3.0
SQLAlchemy==1.3.22
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
termcolor==2.0.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tqdm==4.64.1
twilio==6.50.1
typeguard==2.13.3
types-cryptography==3.3.23
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
